requirements.txt
visdom
fire
ipdb
torchvision
git+https://github.com/pytorch/tnt.git@master